requirements.txt
spidev==3.6
psutil==5.9.4
smbus2==0.4.2
i2cdevice==0.0.7
numpy==1.16.2
Pillow==5.4.1
